lxc_test: change package name to lxc
Problem: they are different ways of declaring a package name for
testing. Usually the package name of tests are the same as the base
package name (in our case "lxc"). However some libraries (including Go's
std libraries) is using the abbrv: "foo_test" for the package name of
tests. This has the caveat that it's not possible to access to the
package's private identifiers. For more information see:
http://stackoverflow.com/questions/19998250/proper-package-naming-for-testing-in-go-lang

Solution: this fixes #3 by changing the package name from "lxc_test" to
"lxc".