type: zero values for State and BackendStore types
Problem: There is no way one can identify if a type is uninitialized
(a.k.a has zero value) or initalized for State or BackendStore types. It
always will return the type STOPPED (or Btrfs). Also the String() method
never will return the string "<INVALID>".

Solution: Increase the enum just by one. This will make uninitialized
values of BackendStore and State as 0. One can then check easily with:

	store := BackendStore
	if store == 0 {
 		return errors.New("backend is not initalized.")
	}

Also the String() method is going to return the "<INVALID>" string.