lxc-binding: add binding for c->attach()
Add a wrapper to call low-level c->attach() which stores the PID of the attached
process in one of its arguments. In contrast to all the other wrappers
go_lxc_attach_no_wait() will not wait on the attached process. This will allow
more flexibility in LXD when executing commands in a container.

Signed-off-by: Christian Brauner <christian.brauner@canonical.com>