lxc-binding: detect lxc_groups_t support through macro
In [1] we're adding macros for all enums. The LXC_ATTACH_SETGROUPS enum
guards the lxc_groups_t feature. So when we also have a define for
LXC_ATTACH_SETGROUPS in addition to the enum we know LXC will have
lxc_groups_t defined and we don't need to do it.

[1]: https://github.com/lxc/lxc/pull/3705
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>